+++ src/main/java/io/mixeway/rest/admin/model/CronSettings.java
@@ -1,0 +1,13 @@
+package io.mixeway.rest.admin.model;
+
+public class CronSettings {
+    private String expresion;
+
+    public String getExpresion() {
+        return expresion;
+    }
+
+    public void setExpresion(String expresion) {
+        this.expresion = expresion;
+    }
+}