+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="0.91"></a>
+## 0.9 (2019-12-29)
+
+### Info
+
+* Fixed bug with Network scan request API
+* Added possibility to put CRON expresion to DB and load if from DB
+
 <a name="0.9"></a>
 ## 0.9 (2019-12-07)